agent/requirements.txt
# AI Investment Analysis Agent

fastapi>=0.109.0
uvicorn>=0.27.0

# LangGraph for orchestration
langchain>=0.1.0
langgraph>=0.0.20
langchain-google-genai>=0.0.6

# Data handling
pydantic>=2.0.0
httpx>=0.26.0

# Environment
python-dotenv>=1.0.0
